pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warc-tools"
version = "0.0.0"
description = "Tools for extracting ETHZ webarchive WARC data and indexing JSONL into Elasticsearch."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  # extractor
  "warcio",
  "beautifulsoup4",
  "pdfplumber",
  "pandas",
  "openpyxl",
  "python-dotenv",

  # indexer
  "elasticsearch",
  "llama-index-core",
  "llama-index-embeddings-ollama",
  "llama-index-embeddings-openai",
  "llama-index-embeddings-huggingface",
  "llama-index-vector-stores-elasticsearch",
]

[project.scripts]
# CLI for extractor: warc to JSONL
warc-extract = "warc_tools.extractor.cli:main"
# CLI for indexer: JSONL to Elasticsearch
warc-index = "warc_tools.indexer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
